+++ examples/hubbard.py
@@ -2,28 +2,25 @@ from ase.units import Ha
 from hotcent.atomic_dft import AtomicDFT
 from hotcent.confinement import PowerConfinement
 
-energies = {}
-for occ2p, label in zip([2, 1, 3], ['neutral', 'cation', 'anion']):
-    atom = AtomicDFT('Si',
-                     confinement=PowerConfinement(r0=40., s=4),
-                     configuration='[Ne] 3s2 3p%d' % occ2p,
-                     valence=['3s', '3p'],
-                     scalarrel=False,
-                     )
-    atom.run()
-    energies[label] = atom.total_energy
+atom = AtomicDFT('C',
+                 xcname='LDA',
+                 confinement=PowerConfinement(r0=40., s=4),
+                 configuration='[He] 2s2 2p2',
+                 valence=['2s', '2p'],
+                 scalarrel=False,
+                 )
 
-EA = energies['neutral'] - energies['anion']
-IE = energies['cation'] - energies['neutral']
-U = IE - EA
+values = []
+schemes = ['central', 'forward', 'backward']
+for scheme in ['central', 'forward', 'backward']:
+    u = atom.get_hubbard_value('2p', scheme=scheme, maxstep=1.)
+    values.append(u)
 
-labels = ['EA', 'IE', 'U']
-values = [EA, IE, U]
-references = [0.044065, 0.288307, 0.244242]
+references = [0.35665217698348783, 0.2694950932099687, 0.4310705719360044]
 eps = 1e-4
 
-print('==========================================')
-print('# Property\t[Ha]\t\t[eV]')
-for l, v, r in zip(labels, values, references):
-    print('%s\t\t%.6f\t%.3f' % (l, v, v * Ha))
-    assert abs(v - r) < eps, (l, v, r, eps) 
+print('\n========================================================')
+print('# Scheme\tU_ref [Ha]\tU [Ha]\t\tU [eV]')
+for s, v, r in zip(schemes, values, references):
+    print('%s\t%.6f\t%.6f\t%.3f' % (s.ljust(8), r, v, v * Ha))
+    assert abs(v - r) < eps, (s, v, r, eps)